+++ packages/pds/src/auth-routes.ts
@@ -15,6 +15,13 @@ export const createRouter = ({ authProvider, cfg }: AppContext): Router => {
       resource_documentation: 'https://atproto.com',
     })
 
+  if (
+    !cfg.service.devMode &&
+    !oauthProtectedResourceMetadata.resource.startsWith('https://')
+  ) {
+    throw new Error('Resource URL must use the https scheme')
+  }
+
   router.get('/.well-known/oauth-protected-resource', (req, res) => {
     res.setHeader('Access-Control-Allow-Origin', '*')
     res.setHeader('Access-Control-Allow-Method', '*')